added database seeding
